+++ lib/Capabilities.php
@@ -68,8 +68,10 @@ class Capabilities implements IPublicCapability {
 				'last-room-activity',
 				'no-ping',
 				'system-messages',
+				'delete-messages',
 				'mention-flag',
 				'in-call-flags',
+				'conversation-call-flags',
 				'notification-levels',
 				'invite-groups-and-mails',
 				'locked-one-to-one-rooms',
@@ -87,6 +89,7 @@ class Capabilities implements IPublicCapability {
 				'phonebook-search',
 				'raise-hand',
 				'room-description',
+				'rich-object-sharing',
 			],
 			'config' => [
 				'attachments' => [