+++ src/components/MessagesList/MessagesList.vue
@@ -131,6 +131,8 @@ export default {
 			previousScrollTopValue: null,
 
 			pollingErrorTimeout: 1,
+
+			oldMessagesPromise: null,
 		}
 	},
 
@@ -162,6 +164,10 @@ export default {
 			const groups = []
 			let lastMessage = null
 			for (const message of this.messagesList) {
+				if (message.systemMessage === 'message_deleted') {
+					continue
+				}
+
 				if (!this.messagesShouldBeGrouped(message, lastMessage)) {
 					// Add the date separator for different days
 					if (this.messagesHaveDifferentDate(message, lastMessage)) {
@@ -425,7 +431,8 @@ export default {
 
 			// Make the request
 			try {
-				const messages = await request({ token, lastKnownMessageId, includeLastKnown: includeLastKnown ? '1' : '0' })
+				this.oldMessagesPromise = request({ token, lastKnownMessageId, includeLastKnown: includeLastKnown ? '1' : '0' })
+				const messages = await this.oldMessagesPromise
 				// Process each messages and adds it to the store
 				messages.data.ocs.data.forEach(message => {
 					if (message.actorType === 'guests') {
@@ -452,10 +459,12 @@ export default {
 						id: newestKnownMessageId,
 					})
 				}
+				this.oldMessagesPromise = null
 			} catch (exception) {
 				if (Axios.isCancel(exception)) {
 					console.debug('The request has been canceled', exception)
 				}
+				this.oldMessagesPromise = null
 			}
 		},
 
@@ -559,6 +568,10 @@ export default {
 				this.displayMessagesLoader = false
 				this.previousScrollTopValue = scrollTop
 			} else if (scrollHeight > elementHeight && scrollTop < 800 && scrollTop <= this.previousScrollTopValue) {
+				if (this.oldMessagesPromise) {
+					// already loading, don't do it twice
+					return
+				}
 				if (scrollTop === 0) {
 					this.displayMessagesLoader = true
 				}